pyproject.toml
[project]
name = "quran-page-spliter"
version = "0.1.0"
description = "A small web application and CLI tool for processing Mushaf (Quran) pages, intelligently crop the page bounds, automatically detect text lines, classify Sura name decorations, and split lines into individual aya segments."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.11.0",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.3",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
